add EXTREME difficulty and BORING flag
EXTREME difficulty is between V_HARD and STUPID.
BORING flag is used when the trick itself is not difficult but very time consuming (e.g. 0% cocoa cave bomb spot) or when influence of luck is significant (e.g. mana wager slide bonk).

This commit is based on backup data as of April 20, 2024